Rename nox.session to nox.sessions.
This paves the way for a @nox.session decorator to
preserve session order.
